fix(bgl/webgpu): silence MSVC C4365 sign conversions
MSVC treats signed/unsigned narrowing as an error where clang stays quiet:
components.size() (size_t) into SlangInt, and the stencil-mask ternaries
whose int literal branch widened into uint32_t. Cast both explicitly.

+++ libs/bgl/src/webgpu/pipeline/RenderPipeline_wgpu.cpp
@@ -29,10 +29,12 @@ namespace bgl
 				return out;
 			};
 
-			result.stencilFront     = stencilFace(state.frontFaceStencil);
-			result.stencilBack      = stencilFace(state.backFaceStencil);
-			result.stencilReadMask  = state.stencilEnable ? state.stencilReadMask : 0;
-			result.stencilWriteMask = state.stencilEnable ? state.stencilWriteMask : 0;
+			result.stencilFront = stencilFace(state.frontFaceStencil);
+			result.stencilBack  = stencilFace(state.backFaceStencil);
+			result.stencilReadMask =
+				state.stencilEnable ? static_cast<uint32_t>(state.stencilReadMask) : 0u;
+			result.stencilWriteMask =
+				state.stencilEnable ? static_cast<uint32_t>(state.stencilWriteMask) : 0u;
 
 			return result;
 		}
